FIX #16125 TIME 4:10 Migrations sql + query params profile + order

+++ migration/21.03/21.03.sql
@@ -65,11 +65,14 @@ INSERT INTO tiles (user_id, type, view, position, color, parameters)
 SELECT id, 'followedMail', 'chart', 0, '#90caf9', '{"chartMode": "status", "chartType": "pie"}' FROM users WHERE status != 'DEL';
 
 
-ALTER TABLE contacts_groups DROP COLUMN IF EXISTS entities;
-ALTER TABLE contacts_groups ADD COLUMN entities jsonb NOT NULL DEFAULT '{}';
-//TODO migration
--- ALTER TABLE contacts_groups DROP COLUMN IF EXISTS public;
-
+DO $$ BEGIN
+    IF (SELECT count(column_name) from information_schema.columns where table_name = 'contacts_groups' and column_name = 'public') THEN
+        ALTER TABLE contacts_groups DROP COLUMN IF EXISTS entities;
+        ALTER TABLE contacts_groups ADD COLUMN entities jsonb NOT NULL DEFAULT '{}';
+        UPDATE contacts_groups SET entities = (select to_jsonb(array_agg(id)) from entities) WHERE public = true;
+        ALTER TABLE contacts_groups DROP COLUMN IF EXISTS public;
+    END IF;
+END$$;
 DO $$ BEGIN
     IF (SELECT count(column_name) from information_schema.columns where table_name = 'contacts_groups_lists' and column_name = 'contact_id') THEN
         ALTER TABLE contacts_groups_lists RENAME COLUMN contact_id TO correspondent_id;
